Module reactive::continuation

A continuation represents some work to be done in the future.

Continuations are also what are stored in the runtime(s) and to execute a continuation we need to call it with some value. This construction is quite similar to the one used by ReactiveML according to this article.

Structs

Map
A continuation that applies a function before calling another continuation.

Pause
A continuation that calls another continuation in the next instant.

Traits

Continuation
A reactive continuation awaiting a value of type V. For the sake of simplicity, continuation must be valid on the static lifetime.

ContinuationPl
Continuation which can be safely passed and shared between different threads. Used for the parallel implementation of the library. The Sync trait is only necessary when signals come into the scene.

ContinuationSt
Continuation to be run in a single thread.