Trait reactive::continuation::Continuation

pub trait Continuation<R, V>: 'static where
    R: Runtime, {
    fn call(self, runtime: &mut R, value: V);

    fn call_box(self: Box<Self>, runtime: &mut R, value: V);

    fn map<F, V2>(self, map: F) -> Map<Self, F>
    where
        Self: Sized,
        F: FnOnce(V2) -> V + 'static,
    { ... }

    fn pause(self) -> Pause<Self>
    where
        Self: Sized,
    { ... }
}

A reactive continuation awaiting a value of type V. For the sake of simplicity, continuation must be valid on the static lifetime.

Required Methods

fn call(self, runtime: &mut R, value: V)

Calls the continuation.

fn call_box(self: Box<Self>, runtime: &mut R, value: V)

Calls the continuation. Works even if the continuation is boxed.

This is necessary because the size of a value must be known to unbox it. It is thus impossible to take the ownership of a Box<Continuation> whitout knowing the underlying type of the Continuation.

Provided Methods

fn map<F, V2>(self, map: F) -> Map<Self, F> where
Self: Sized,
F: FnOnce(V2) -> V + 'static,

Creates a new continuation that applies a function to the input value before calling Self.

fn pause(self) -> Pause<Self> where
Self: Sized,

Create a new continuation that calls Self in the next instant.