Trait reactive::continuation::ContinuationPl

pub trait ContinuationPl<V>: Continuation<ParallelRuntime, V> + Send + Sync { }

Continuation which can be safely passed and shared between different threads. Used for the parallel implementation of the library. The Sync trait is only necessary when signals come into the scene.