Module reactive::process

A process encapsulates what should be runned by a reactive engine.

Starting from the function value_proc or some signal, users are free to define their own process to be executed in the reactive environment using the methods that are offered by the traits Process and ProcessMut. There is no need to manipulate directly the runtime engines because we only need to call execute_process, execute_process_parallel, or execute_process_parallel_with_main at the end to execute the process.

If a process is only defined with things found in this module (in other words, no signal is used), we can execute it in the two kinds of runtime as long as it's Send and Sync.

Concerning the code structure, the code for the parallel and non-parallel part of the library are in fact very similar. However, I'm not able to figure out a way to design traits so that all can be put together through some abstraction in order to reduce repeated code.

The problem is that the method call has a generic type parameter C and in the two cases it must implement the trait ContinuationSt or ContinuationPl. I didn't find a way to integrate this information into a same trait, so I must have two separated trait implementations. For example, a trait that can able to parametrized by another trait could be very helpful, but that doesn't exist in Rust at this moment (and I admit that I don't even know if this is possible from a theretical viewpoint).

Structs

AndThen
Chains a computation onto the end of another process.

Flatten
Flattens the process when it returns another process to get only the result of the final process.

IfElse
Selects the process to run according to what the previous process returns.

Join
Parallel composition of two processes.

JoinAll
A process which takes a list of processes and terminates with a vector of the completed values.

Loop
Repeats a process forever.

Map
A process that applies a function to the returned value of another process.

Pause
The process is suspended until next instant.

Repeat
Repeats a process several times and collects all the produced values in a vector which is returned at the end of the loop.

Then
Executes the second process while ignoring the returned value of the first process.

Value
A process that returns a value of type V.

While
Repeats a process having LoopStatus as return type until it returns Exit(v) in which case the created process exits and returns v.

Enums

LoopStatus
Indicates if a loop is terminated or not.

Traits

ConstraintOnValue
This is a workaround to have constraints on associated types. Must be implemented by every type that want to represent some parallel process.

Process
A abstract reactive process. A method call is in fact also necessary. Please see ProcessSt and ProcessPl for more information.

ProcessMut
A process that can be executed multiple times, modifying its environement each time.

ProcessMutPl
A repeatable reactive process that can be safely passed and shared between threads.

ProcessMutSt
A repeatable reactive process to be executed in a single thread.

ProcessPl
A reactive process that can be safely passed and shared between threads.

ProcessSt
A reactive process to be executed in a single thread.

Functions

execute_process
Executes a process in a newly created runtime and return its value (without parallization).

execute_process_parallel
Executes a process in newly created runtimes and return its value (with parallization).

execute_process_parallel_with_main
Executes a process in newly created runtimes and return its value. Each runtime is runned in a separated child thread and these threads are runned in parallel with a main function that is executed in the main thread. This construction is necessary when some part of the program must be executed in the main thread.

join_all
Creates a process that executes a collection of processes in parallel and collects the results into a destination Vec<T> in the same order as they were provided.
However, since the Rust's type system force all the values in a vector to take the same type and most of the processes have different type from each other, this may not be very useful. (What's worse, a process cannot be made into a trait object.)

value_proc
Create a new process that returns the value v immediately.