Function reactive::process::execute_process_parallel_with_main

pub fn execute_process_parallel_with_main<F, P>(
    f: F, 
    p: P, 
    num_runtimes: usize
) -> P::Value where
    F: FnOnce(),
    P: ProcessPl, 

Executes a process in newly created runtimes and return its value. Each runtime is runned in a separated child thread and these threads are runned in parallel with a main function that is executed in the main thread. This construction is necessary when some part of the program must be executed in the main thread.