Struct reactive::process::Repeat

pub struct Repeat<P> { /* fields omitted */ }

Repeats a process several times and collects all the produced values in a vector which is returned at the end of the loop.