Trait reactive::process::ConstraintOnValue

pub trait ConstraintOnValue {
    type T: Send + Sync;
}

This is a workaround to have constraints on associated types. Must be implemented by every type that want to represent some parallel process.

Associated Types

type T: Send + Sync