Trait reactive::process::Process

pub trait Process: 'static {
    type Value;
    fn pause(self) -> Pause<Self>
    where
        Self: Sized,
    { ... }

    fn map<F, V>(self, map: F) -> Map<Self, F>
    where
        Self: Sized,
        F: FnOnce(Self::Value) -> V + 'static,
    { ... }

    fn flatten(self) -> Flatten<Self>
    where
        Self: Sized,
        Self::Value: Process,
    { ... }

    fn and_then<F, P>(self, chain: F) -> AndThen<Self, F>
    where
        Self: Sized,
        F: FnOnce(Self::Value) -> P + 'static,
        P: Process,
    { ... }

    fn then<P>(self, successor: P) -> Then<Self, P>
    where
        Self: Sized,
        P: Process,
    { ... }

    fn if_else<P1, P2, V>(
        self, 
        if_branch: P1, 
        else_branch: P2
    ) -> IfElse<Self, P1, P2>
    where
        Self: Process<Value = bool> + Sized,
        P1: Process<Value = V>,
        P2: Process<Value = V>,
    { ... }

    fn join<P>(self, proc2: P) -> Join<Self, P>
    where
        Self: Sized,
        P: Process,
    { ... }
}

A abstract reactive process. A method call is in fact also necessary. Please see ProcessSt and ProcessPl for more information.

Associated Types

type Value

The value created by the process.

Provided Methods

fn pause(self) -> Pause<Self> where
Self: Sized,

Suspends the execution of a process until next instant.

fn map<F, V>(self, map: F) -> Map<Self, F> where
Self: Sized,
F: FnOnce(Self::Value) -> V + 'static,

Applies a function to the value returned by the process before passing it to its continuation.

fn flatten(self) -> Flatten<Self> where
Self: Sized,
Self::Value: Process,

Flattens the execution of a process when its returned value is itself another process.

fn and_then<F, P>(self, chain: F) -> AndThen<Self, F> where
Self: Sized,
F: FnOnce(Self::Value) -> P + 'static,
P: Process,

Chains another process after the exectution of one process (like the bind for a monad).

fn then<P>(self, successor: P) -> Then<Self, P> where
Self: Sized,
P: Process,

Executes a second process after one process terminates. The returned value of the first process is ignored.

fn if_else<P1, P2, V>(
self,
if_branch: P1,
else_branch: P2
) -> IfElse<Self, P1, P2> where
Self: Process<Value = bool> + Sized,
P1: Process<Value = V>,
P2: Process<Value = V>,

Decides whether to execute if_branch or else_branch according to the returned value of self, which must be of type bool.
The combinator and_then defined earlier together with the built-in if-else branching in Rust cannot allow us to achieve the same purpose since if branch and else branch in Rust must result in the same type.

fn join<P>(self, proc2: P) -> Join<Self, P> where
Self: Sized,
P: Process,

Executes two processes in parallel.