Trait reactive::process::ProcessMut

pub trait ProcessMut: Process {
    fn loop_proc(self) -> Loop<Self>
    where
        Self: Sized,
    { ... }

    fn repeat(self, times: usize) -> Repeat<Self>
    where
        Self: Sized,
    { ... }

    fn while_proc<V>(self) -> While<Self>
    where
        Self: ProcessMut<Value = LoopStatus<V>> + Sized,
    { ... }
}

A process that can be executed multiple times, modifying its environement each time.

Provided Methods

fn loop_proc(self) -> Loop<Self> where
Self: Sized,

An infinite loop. The returned value of the process is ignored if any.

fn repeat(self, times: usize) -> Repeat<Self> where
Self: Sized,

A simple loop that just repeats the process a given number of times and collects all the returned values in a vector.

fn while_proc<V>(self) -> While<Self> where
Self: ProcessMut<Value = LoopStatus<V>> + Sized,

A classic loop that continues or stops accroding to the returned value of the process.