Trait reactive::process::ProcessMutPl

pub trait ProcessMutPl: ProcessMut + ProcessPl {
    fn call_mut<C>(self, runtime: &mut ParallelRuntime, next: C)
    where
        Self: Sized,
        C: ContinuationPl<(Self, Self::Value)>;
}

A repeatable reactive process that can be safely passed and shared between threads.

Required Methods

fn call_mut<C>(self, runtime: &mut ParallelRuntime, next: C) where
Self: Sized,
C: ContinuationPl<(Self, Self::Value)>,

Executes the mutable process in the runtime, then calls next with the process and the process's return value.