Trait reactive::process::ProcessPl

pub trait ProcessPl: Process<Value = Self::T> + ConstraintOnValue + Send + Sync {
    fn call<C>(self, runtime: &mut ParallelRuntime, next: C)
    where
        C: ContinuationPl<Self::Value>;
}

A reactive process that can be safely passed and shared between threads.

Required Methods

fn call<C>(self, runtime: &mut ParallelRuntime, next: C) where
C: ContinuationPl<Self::Value>,

Executes the reactive process in the runtime, calls next with the resulting value.