Trait reactive::process::ProcessSt

pub trait ProcessSt: Process {
    fn call<C>(self, runtime: &mut SingleThreadRuntime, next: C)
    where
        C: ContinuationSt<Self::Value>;
}

A reactive process to be executed in a single thread.

Required Methods

fn call<C>(self, runtime: &mut SingleThreadRuntime, next: C) where
C: ContinuationSt<Self::Value>,

Executes the reactive process in the runtime, calls next with the resulting value.