Module reactive::runtime

One runtime have some tasks to do at hand and chooses one to do at each time.

This is where the execution engines are defined. The tasks to be carry out are some continuations which are charaterized by the trait continuation::Continuation. Each instants terminates only when there are no more tasks to do at this instant. The whole execution terminates when no more continuation is left in the runtime(s) and no signal is awaiting for emisssion.

A SingleThreadRuntime is itself the whole execution engine and is runned on the main thread. In contrast, a ParallelRuntime is spawned on a new thread and is only one part of the whole parallel engine which is given by the struct ParallelRuntimeCollection.

Structs

ParallelRuntime
Runtime for executing reactive continuations in a separated thread.

ParallelRuntimeCollection
Collect multiple ParallelRuntime to form a parallel execution engine.

SingleThreadRuntime
Runtime for executing reactive continuations in the main thread.

Traits

Runtime
Must be implemented by all concrete runtime types.