Struct reactive::runtime::ParallelRuntime

pub struct ParallelRuntime { /* fields omitted */ }

Runtime for executing reactive continuations in a separated thread.

Every runtime has its owned FIFO containing continuations to be carried out and to ensure thread work balance a work stealing strategy is adopted. To avoid stealing from a thread that is not working all the runtimes share a common working pool which indicates the working runtimes. There are two synchronization using barrier at the end of each instant, one for the main works and one for the works to be executed only at the end of instant (there are notably the await constructions of multi-producer signals). A conditional variable is then used to know if the whole engine should terminate or not.