Struct reactive::runtime::ParallelRuntimeCollection

pub struct ParallelRuntimeCollection { /* fields omitted */ }

Collect multiple ParallelRuntime to form a parallel execution engine.

Methods

impl ParallelRuntimeCollection

fn new(num_runtimes: usize) -> Self

Creates a new parallel execution engine containing a certain number of runtimes. Each runtime is runned in a separeted thread.

fn execute<F>(&mut self, main_thread_function: F) where
F: FnOnce(),

Executes in parallel all the runtimes contained in the collection, with one thread for each runtime.
When the library is compiled with the feature abort_if_panic, the process that is executing the program is aborted whenever a child thread panics. Maybe there's another way to smartly deal with panics coming from child threads but I am not particularly working on it.

fn register_work(&mut self, c: Box<ContinuationPl<()>>)

Adds some work to be executed by the execution engine.