Module reactive::signal

Inter-process communication relies on the use of signals.

Four kinds of signals are defined: PureSignal, MpmcSignal, MpscSignal and SpmcSignal. Except for PureSignal, all signals are emitted with some value and store some value which can be communicated.

We should notice that the implementaion for MpscSignal and SpmcSignal may not be very satisfactory. We can imagine having some sender or receiver for a signal and it is consumed once used. However, this means that the signal can only be emitted or awaited in one place in the code, and this is not what I look for. The goal is to force the signal to be emitted or consumed only once at each instant, but not only once in the whole program. Since I have no idea how this can be down at compile time, what I finally did is to check this dynamically. The program panics when some undesired behavior is detected.

On the other hand, the signals used for the non-parallel and the parallel version of the library are different, so the user must decide which sort of signal (the parallel or non-parallel ones) to use from the beginning. I would love to have something like in the case of processes: the real behavior of the signal is only determined when it's associated with some particular runtime, but I didn't find a way to do this.

Reexports

pub use self::pure_signal::PureSignal;

pub use self::valued_signal::ValuedSignal;

pub use self::parallel::*;

pub use self::single_thread::*;

Modules

parallel
Signals to use for parallel reactive engine.

pure_signal
A reactive signal without value.

single_thread
Signals to use for non-parallel reactive engine.

valued_signal
A reactive signal with value.

Structs

AwaitImmediate
Process that awaits the emission of a signal.

PresentElse
Process that tests the status of a signal and chooses the branch to execute according to the result.

Traits

Signal
A reactive signal.