Trait reactive::signal::Signal

pub trait Signal: Clone + 'static {
    type RuntimeRef;
    fn runtime(&self) -> Self::RuntimeRef;

    fn await_immediate(&self) -> AwaitImmediate<Self>
    where
        Self: Sized,
    { ... }

    fn present_else<P1, P2, V>(
        &self, 
        p1: P1, 
        p2: P2
    ) -> PresentElse<Self, P1, P2>
    where
        Self: Sized,
        P1: Process<Value = V>,
        P2: Process<Value = V>,
    { ... }
}

A reactive signal.
The signal implement the trait Clone to assure that it can be used multiple times in the program. However, note that for most of the constructions clone is used implicitly so one can pass the signal directly. The user needs to call clone explicitly only in scenarios where the signal's ownership must be shared in different places, ex: closure.

Associated Types

type RuntimeRef

The runtime reference type associated with the signal.

Required Methods

fn runtime(&self) -> Self::RuntimeRef

Returns a reference to the signal's runtime.

Provided Methods

fn await_immediate(&self) -> AwaitImmediate<Self> where
Self: Sized,

Returns a process that waits for the next emission of the signal, current instant included.

fn present_else<P1, P2, V>(&self, p1: P1, p2: P2) -> PresentElse<Self, P1, P2> where
Self: Sized,
P1: Process<Value = V>,
P2: Process<Value = V>,

Test the status of a signal s. If the signal is present, the process p1 is executed instantaneously, other wise p2 is executed at the following instant.