Module reactive::signal::valued_signal

Structs

AwaitValue
Process awaiting a signal to be emitted to fetch its value.

EmitValue
Process that represents an emission of a signal with some value.

MpSignal
For multi-producer signal, await terminates at the following instant.

SpSignal
For single-producer signal, await terminates immediately.

TryEmitValue
Process that represents an emission of a signal with some value.

Traits

CanEmit

GetValue

SignalType
Define some subtypes that a signal with value can have.

ValuedSignal
Defines the behavior of a pure signal. This is the interface exposed to users.