Trait reactive::signal::valued_signal::ValuedSignal

pub trait ValuedSignal: Signal {
    type Stored;
    type SigType: SignalType;
    fn emit<A>(&self, emitted: A) -> EmitValue<Self, A>
    where
        Self: Sized,
    { ... }

    fn await(&self) -> AwaitValue<Self, Self::SigType>
    where
        Self: Sized,
    { ... }
}

A reactive signal with value.

Associated Types

type Stored

The value stored in the signal.

type SigType: SignalType

A more specific type of a signal. This is necessary because some generic implementation may depend on the specific type of the signal.

Provided Methods

fn emit<A>(&self, emitted: A) -> EmitValue<Self, A> where
Self: Sized,

Returns a process that emits the signal with value emitted when it is called.

fn await(&self) -> AwaitValue<Self, Self::SigType> where
Self: Sized,

Waits the signal to be emitted and gets its content.

For a single-producer signal the process terminates immediately and for a multi-producer signal it terminates at the following instant. For example, when s1 is a spmc signal and when s2 is a mpmc signal, s1.await().pause() is semantically equivalent to s2.await().